Reduce instances of em dash constructions
Em dash constructions are meant for making very complicated information more legible. In most other instances, bullets and colons are preferred. I still kept properties in bold, because they appear to the user in the UI.

I also changed an instance of modeler to app, as we do not use the term modeler anymore.